Working on a generic list data structure.
The list will be used for, among other things, instructions. The list
is fairly strict in its use of value semantics, so at low/no
optimization levels, it is likely to perform poorly, but with
optimization it should fare well. This commit is mostly to test how the
kernel handles templates that appear in the .S generated files.