Solved issue (on Mac)
So the main issue was that the mac heap is over 4gb away from code. A
call that was being encoded at static init time (.init section) was
failing, and it wasn't and issue in GDB. I changed how static
initialization of code happens; it is now added to the init list, and
initialization of code is done with granary::init(). Also, for issues
that don't present themselves in gdb, I have added a segfault signal
handler, which enters an infinite loop, giving the programmer ample
time to attach gdb to the process to debug.