Enabled user-space specific mangling for far memory operands.
In user space (at least on Mac OS X), %rip-relative memory operands
tend to be > 4GB away from the code cache. These changes specifically
look for such operands and attempt to fix them directly by spilling the
computed address of the %rip-relative operand to a register, then using
that register as a base/disp operand.

A bunch of the functionality in place for memory operands is also going
to be used for handling the similar case of direct CTIs that need to be
mangled to use a vtable entry.