Starting work on building the hash table.
Decided to use an enum instead of macros for detach ids. Now, functions
will not be associated with their actual in-memory addresses (not
viable in userspace; and presents a security risk in kernel space);
instead, each function will have a unique id (implicitly handled by
being an enum constant) and that id will be an index into a table where
the actual address is stored (by virtue of run time linking). The
entries of the hash table will then be this static table. Thus, the
hash function will map an address to its id.