Speed up retrieving module sections from target
By replacing the small bit of shell code to retrieve the module
section-per-symbol with pure Python, we save thousands of subprocesses
invocations which dramatically speeds up the process of remote loading.
We actually end up copying the remoteload.py script itself to the target
to achieve this, which is a neat turn around.